Be consistent checking c standard in assert.h (#527)
Assert header is split into two to handle the idempotency since
commit b3260c2b "17491 assert(3c) prototype visibility". The macro
is in one and the prototype for the internal function in the other.
There are different variants for pre-c99 and c99 and the macro and
prototype needs to match. The aforementioned commit changed the
prototype to use "_STDC_C99" while the macro is using
"__STDC_VERSION__".

Normally _STDC_C99 and __STDC_VERSION__ would match and this wouldn't be
a problem. As sys/feature_tests.h sets _STD_C99 based on
__STDC_VERSION__ value.

However the gcc10 compiler is patched to set _STD_C99 as a predefined
macro for C++ code.
https://github.com/TritonDataCenter/pkgsrc-extra/commit/e295cdbf14a6991ff91ad6e5624ddec0f96244ca

This caused using assert from c++ code to fail because the macro
expanded a function that there was no proto type for.